requirements.txt
Pillow
PyAutoGUI
Numpy
tensorflow==2.3.1